requirements.txt
# Core dependencies
python-dotenv>=0.19.0
langchain>=0.0.300
langchain-openai>=0.0.5
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.28.0

# API dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=22.3.0
isort>=5.10.1
flake8>=4.0.1
mypy>=0.990

# LLM providers
bytez>=0.0.0
